Streamlit/requirements.txt
streamlit
google.generativeai
youtube_search
streamlit_extras
pytest-shutil
#tempfile
opencv-python-headless
datetime
#altair==4.0